Greeting 删除操作前，检查传入的id是否存在如果不存在在返回

+++ src/main/java/com/euky/ws/service/GreetingServiceBean.java
@@ -51,6 +51,10 @@ public class GreetingServiceBean implements  GreetingService {
     @Override
     @CacheEvict(value = "greetings", key = "#id")
     public void delete(Long id) {
+        if (findOne(id) == null) {
+            return;
+        }
+
         greetingRepository.delete(id);
     }
